Do cleanup when service process exits
Some service client module, such as kafka-python will buffer data before send to the server,
if for some reason, the worker process is killed and the data still in buffer, this will cause data loss,
so, here we add a hook, that will be executed before the process exits.

NOTE: this is not a 100% solution for data safe, just a improvement.